Added a .jsp page to diplay Profile information. Uses the same scheme as the other .jsp pages with regards to header, footer, and how the buttons are laid out, by that if logged in or not, and what displays as a result of either.
This page would then access the arraylist created, and access the
data within. An if statement was added in case there was no match in the
arraylist for the username supplied, prompting an error message.
If the information needed was there, it would display each bit alongside,
what each pice of information each was (awkward wording)